+++ webots_ros2_tiago/launch/tiago.launch.py
@@ -39,9 +39,22 @@ def generate_launch_description():
                                     node_executable='tiago_controller',
                                     parameters=[{'synchronization': synchronization}],
                                     output='screen')
+
+    # Rviz node
+    use_rviz = launch.substitutions.LaunchConfiguration('rviz', default=False)
+    rviz_config = os.path.join(get_package_share_directory('webots_ros2_tiago'), 'resource', 'odometry.rviz')
+    rviz = launch_ros.actions.Node(
+        package='rviz2',
+        node_executable='rviz2',
+        output='screen',
+        arguments=['--display-config=' + rviz_config],
+        condition=launch.conditions.IfCondition(use_rviz)
+    )
+
     return launch.LaunchDescription([
         webots,
         controller,
+        rviz,
         # Shutdown launch when Webots exits.
         launch.actions.RegisterEventHandler(
             event_handler=launch.event_handlers.OnProcessExit(