+++ webots_ros2_core/CHANGELOG.rst
@@ -8,6 +8,7 @@ Changelog for package webots_ros2_core
 * Added a new tf_publisher node to publish tf of desired Webots nodes.
 * Added the possibility to run nodes in synchronized mode (using the 'synchronization' parameter).
 * Added a laser_publisher module.
+* Added better support for differential drive robots ('WebotsDifferentialDriveNode' class).
 
 0.0.2 (2019-09-23)
 ------------------